// src/app/shared/models/coupon.model.ts
/* eslint-disable @typescript-eslint/no-empty-interface */
/* eslint-disable @typescript-eslint/naming-convention */
export interface ICoupon {
    no?: number;
    id?: string;
    code?: string;
    creation_date?: string;
    description?: boolean;
    discount?: string;
    discount_type?: string;
    enable?: boolean;
    image_url?: string;
    name?: string;
  }